guardrails/output_guardrail/pyproject.toml
[project]
name = "output-guardrail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai-agents>=0.2.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
